+++ Tests/StreamChatTests/APIClient/Endpoints/Payloads/FileUploadPayload_Tests.swift
@@ -10,6 +10,6 @@ final class FileUploadPayload_Tests: XCTestCase {
     func test_payload_isDeserialized() throws {
         let json = XCTestCase.mockData(fromJSONFile: "FileUploadPayload")
         let payload = try JSONDecoder.default.decode(FileUploadPayload.self, from: json)
-        XCTAssertEqual(payload.file, URL(string: "https://i.imgur.com/EgEPqWZ.jpg"))
+        XCTAssertEqual(payload.fileURL, URL(string: "https://i.imgur.com/EgEPqWZ.jpg"))
     }
 }