+++ Tests/StreamChatTests/Models/Attachments/AnyAttachmentUpdater_Tests.swift
@@ -1,0 +1,37 @@
+//
+// Copyright © 2022 Stream.io Inc. All rights reserved.
+//
+
+@testable import StreamChat
+import XCTest
+
+final class AnyAttachmentUpdater_Tests: XCTestCase {
+    let sut = AnyAttachmentUpdater()
+
+    func test_update() throws {
+        var attachment = ChatMessageImageAttachment(
+            id: .init(cid: .unique, messageId: .unique, index: .unique),
+            type: .image,
+            payload: .init(title: "old", imageRemoteURL: .localYodaImage, extraData: [:]),
+            uploadingState: nil
+        ).asAnyAttachment
+
+        let expectation = expectation(description: "should update the attachment payload")
+
+        sut.update(&attachment, forPayload: ImageAttachmentPayload.self) { payload in
+            payload.title = "new"
+            payload.extraData = [
+                "thumbnailUrl": .string("fakeUrl")
+            ]
+
+            expectation.fulfill()
+        }
+
+        waitForExpectations(timeout: 0.5)
+
+        let imageAttachment = attachment.attachment(payloadType: ImageAttachmentPayload.self)
+
+        XCTAssertEqual(imageAttachment?.payload.title, "new")
+        XCTAssertEqual(imageAttachment?.payload.extraData?["thumbnailUrl"]?.stringValue, "fakeUrl")
+    }
+}